+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+## Release 2.0.3
+* refactor(spc): Refactored cloudflare-spc command
+
+
 ## Release 2.0.2
 * improvement: Ask to delete all rules, versus one by one.
 * improvement: Created profile with mwp-rules-v2 including event calendars bot blocking
+* improvement: Created generate-readme.sh to generate/add CHANGELOG.md to README.md
 
 
 ## Release 2.0.1